# contracts/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Swiss Life Triage Agent API",
    "version": "0.1.0"
  },
  "paths": {
...
    "/api/tickets": {
      "get": {
        "tags": [
          "tickets"
        ],
        "summary": "List Tickets",
        "operationId": "list_tickets",
        "parameters": [
          {
            "name": "state",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "enum": [
                    "new",
                    "proposed",
                    "approved",
                    "edited",
                    "rejected"
                  ],
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "State"
            }
          },
          {
            "name": "source",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "enum": [
                    "challenge",
                    "training",
                    "manual",
                    "email"
                  ],
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Source"
            }
          },
          {
            "name": "q",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Search summary and description",
              "title": "Q"
            },
            "description": "Search summary and description"
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 500,
              "minimum": 1,
              "default": 50,
              "title": "Limit"
            }
          },
          {
            "name": "offset",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 0,
              "default": 0,
              "title": "Offset"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TicketPage"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "tickets"
        ],
        "summary": "Create Ticket",
        "operationId": "create_ticket",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TicketCreate"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TicketOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/tickets/from-email": {
      "post": {
        "tags": [
          "tickets"
        ],
        "summary": "Ingest Email",
        "operationId": "ingest_email",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/EmailIngest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TicketOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/tickets/import": {
      "post": {
        "tags": [
          "tickets"
        ],
        "summary": "Import Tickets",
        "operationId": "import_tickets",
        "requestBody": {
          "content": {
            "multipart/form-data": {
              "schema": {
                "$ref": "#/components/schemas/Body_import_tickets"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ImportResult"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/tickets/{ticket_id}": {
      "get": {
        "tags": [
          "tickets"
        ],
        "summary": "Get Ticket",
        "operationId": "get_ticket",
        "parameters": [
          {
            "name": "ticket_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Ticket Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TicketDetail"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "tickets"
        ],
        "summary": "Delete Ticket",
        "operationId": "delete_ticket",
        "parameters": [
          {
            "name": "ticket_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Ticket Id"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/tickets/{ticket_id}/triage": {
      "post": {
        "tags": [
          "triage"
        ],
        "summary": "Triage Ticket",
        "operationId": "triage_ticket",
        "parameters": [
          {
            "name": "ticket_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Ticket Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TriageResultOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/triage/batch": {
      "post": {
        "tags": [
          "triage"
        ],
        "summary": "Triage Batch",
        "operationId": "triage_batch",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/BatchTriageRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BatchTriageResult"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/triage/{result_id}": {
      "get": {
        "tags": [
          "triage"
        ],
        "summary": "Get Triage Result",
        "operationId": "get_triage_result",
        "parameters": [
          {
            "name": "result_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Result Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TriageResultOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/triage/{result_id}/review": {
      "post": {
        "tags": [
          "triage"
        ],
        "summary": "Review Triage",
        "operationId": "review_triage",
        "parameters": [
          {
            "name": "result_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Result Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ReviewCreate"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ReviewOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "BatchTriageRequest": {
        "properties": {
          "ticket_ids": {
            "anyOf": [
              {
                "items": {
                  "type": "string",
                  "format": "uuid"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Ticket Ids",
            "description": "Omit to triage every ticket in state 'new'"
          },
          "source": {
            "anyOf": [
              {
                "type": "string",
                "enum": [
                  "challenge",
                  "training",
                  "manual",
                  "email"
                ]
              },
              {
                "type": "null"
              }
            ],
            "title": "Source"
          }
        },
        "type": "object",
        "title": "BatchTriageRequest"
      },
...
      "Body_import_tickets": {
        "properties": {
          "file": {
            "type": "string",
            "contentMediaType": "application/octet-stream",
            "title": "File",
            "description": "Jira export JSON (challenge or training format)"
          },
          "source": {
            "type": "string",
            "enum": [
              "challenge",
              "training",
              "manual",
              "email"
            ],
            "title": "Source",
            "default": "challenge"
          }
        },
        "type": "object",
        "required": [
          "file"
        ],
        "title": "Body_import_tickets"
      },
...
      "EmailIngest": {
        "properties": {
          "from_address": {
            "type": "string",
            "title": "From Address"
          },
          "subject": {
            "type": "string",
            "title": "Subject"
          },
          "body": {
            "type": "string",
            "title": "Body"
          },
          "business_entity": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Business Entity"
          }
        },
        "type": "object",
        "required": [
          "from_address",
          "subject",
          "body"
        ],
        "title": "EmailIngest"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "ImportResult": {
        "properties": {
          "imported": {
            "type": "integer",
            "title": "Imported"
          },
          "skipped": {
            "type": "integer",
            "title": "Skipped"
          }
        },
        "type": "object",
        "required": [
          "imported",
          "skipped"
        ],
        "title": "ImportResult"
      },
...
      "TicketCreate": {
        "properties": {
          "work_type": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Work Type"
          },
          "request_type": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Request Type"
          },
          "summary": {
            "type": "string",
            "title": "Summary"
          },
          "description": {
            "type": "string",
            "title": "Description",
            "default": ""
          },
          "affected_service": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Affected Service",
            "description": "Intake value; may be wrong"
          },
          "business_entity": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Business Entity"
          },
          "reporter": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Reporter"
          },
          "urgency": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Urgency"
          },
          "impact": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Impact"
          },
          "priority": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Priority"
          },
          "status": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Status",
            "default": "open"
          },
          "linked_issues": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Linked Issues",
            "default": []
          },
          "comments": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Comments",
            "default": []
          },
          "source": {
            "type": "string",
            "enum": [
              "challenge",
              "training",
              "manual",
              "email"
            ],
            "title": "Source",
            "default": "manual"
          }
        },
        "type": "object",
        "required": [
          "summary"
        ],
        "title": "TicketCreate"
      },
      "TicketDetail": {
        "properties": {
          "work_type": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Work Type"
          },
          "request_type": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Request Type"
          },
          "summary": {
            "type": "string",
            "title": "Summary"
          },
          "description": {
            "type": "string",
            "title": "Description",
            "default": ""
          },
          "affected_service": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Affected Service",
            "description": "Intake value; may be wrong"
          },
          "business_entity": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Business Entity"
          },
          "reporter": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Reporter"
          },
          "urgency": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Urgency"
          },
          "impact": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Impact"
          },
          "priority": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Priority"
          },
          "status": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Status",
            "default": "open"
          },
          "linked_issues": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Linked Issues",
            "default": []
          },
          "comments": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Comments",
            "default": []
          },
          "id": {
            "type": "string",
            "format": "uuid",
            "title": "Id"
          },
          "number": {
            "type": "integer",
            "title": "Number"
          },
          "source": {
            "type": "string",
            "enum": [
              "challenge",
              "training",
              "manual",
              "email"
            ],
            "title": "Source"
          },
          "triage_state": {
            "type": "string",
            "enum": [
              "new",
              "proposed",
              "approved",
              "edited",
              "rejected"
            ],
            "title": "Triage State"
          },
          "source_created_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Source Created At"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          },
          "latest_triage": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/TriageResultOut"
              },
              {
                "type": "null"
              }
            ]
          },
          "reviews": {
            "items": {
              "$ref": "#/components/schemas/ReviewOut"
            },
            "type": "array",
            "title": "Reviews"
          }
        },
        "type": "object",
        "required": [
          "summary",
          "id",
          "number",
          "source",
          "triage_state",
          "source_created_at",
          "created_at",
          "latest_triage",
          "reviews"
        ],
        "title": "TicketDetail"
      },
      "TicketOut": {
        "properties": {
          "work_type": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Work Type"
          },
          "request_type": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Request Type"
          },
          "summary": {
            "type": "string",
            "title": "Summary"
          },
          "description": {
            "type": "string",
            "title": "Description",
            "default": ""
          },
          "affected_service": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Affected Service",
            "description": "Intake value; may be wrong"
          },
          "business_entity": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Business Entity"
          },
          "reporter": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Reporter"
          },
          "urgency": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Urgency"
          },
          "impact": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Impact"
          },
          "priority": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Priority"
          },
          "status": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Status",
            "default": "open"
          },
          "linked_issues": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Linked Issues",
            "default": []
          },
          "comments": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Comments",
            "default": []
          },
          "id": {
            "type": "string",
            "format": "uuid",
            "title": "Id"
          },
          "number": {
            "type": "integer",
            "title": "Number"
          },
          "source": {
            "type": "string",
            "enum": [
              "challenge",
              "training",
              "manual",
              "email"
            ],
            "title": "Source"
          },
          "triage_state": {
            "type": "string",
            "enum": [
              "new",
              "proposed",
              "approved",
              "edited",
              "rejected"
            ],
            "title": "Triage State"
          },
          "source_created_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Source Created At"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          }
        },
        "type": "object",
        "required": [
          "summary",
          "id",
          "number",
          "source",
          "triage_state",
          "source_created_at",
          "created_at"
        ],
        "title": "TicketOut"
      },
      "TicketPage": {
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/TicketOut"
            },
            "type": "array",
            "title": "Items"
          },
          "total": {
            "type": "integer",
            "title": "Total"
          }
        },
        "type": "object",
        "required": [
          "items",
          "total"
        ],
        "title": "TicketPage"
      },
...
    }
  }
}